# Ref/optimize init