--- a/src/did-provider.ts
+++ b/src/did-provider.ts
@@ -41,34 +41,75 @@ interface AuthParams {
   paths: Array<string>
 }
 
+interface JWSSignature {
+  protected: string
+  signature: string
+}
+
+interface GeneralJWS {
+  payload: string
+  signatures: Array<JWSSignature>
+}
+
+function toGeneralJWS(jws: string): GeneralJWS {
+  const [protectedHeader, payload, signature] = jws.split('.')
+  return {
+    payload,
+    signatures: [{ protected: protectedHeader, signature }],
+  }
+}
+
+async function sign(
+  payload: Record<string, any>,
+  didWithFragment: string,
+  keyring: Keyring,
+  threeIdx: ThreeIDX,
+  protectedHeader: Record<string, any> = {},
+  revocable?: boolean
+): string {
+  let [did, keyFragment] = didWithFragment.split('#') // eslint-disable-line prefer-const
+  let kid, signer
+  if (did.startsWith('did:key:')) {
+    const pubkey = did.split(':')[2]
+    kid = `${did}#${pubkey}`
+    signer = keyring.getMgmtSigner(pubkey)
+  } else {
+    if (did !== threeIdx.id) throw new Error(`Unknown DID: ${did}`)
+    const version = await threeIdx.get3idVersion()
+    if (!keyFragment) keyFragment = keyring.getKeyFragment(version)
+    kid = `${did}${revocable ? '' : `?version-id=${version}`}#${keyFragment}`
+    signer = keyring.getSigner(version)
+  }
+  const header = toStableObject(Object.assign(protectedHeader, { kid }))
+  return createJWS(toStableObject(payload), signer, header)
+}
+
 export const didMethods: HandlerMethods<Context> = {
-  did_authenticate: async ({ permissions, threeIdx, origin, forcedDID }, params: AuthParams) => {
+  did_authenticate: async ({ permissions, keyring, threeIdx, origin, forcedDID }, params: AuthParams) => {
     const paths = await permissions.request(origin, params.paths || [])
     // paths should be an array if permission granted
     // may be a subset or requested paths or empty array
     if (paths === null) throw new RPCError(4001, 'User Rejected Request')
-    return { did: forcedDID || threeIdx.id, paths }
+    const response = await sign(
+      {
+        did: forcedDID || threeIdx.id,
+        aud: params.aud,
+        nonce: params.nonce,
+        paths,
+        exp: Math.floor(Date.now() / 1000) + 600, // expires 10 min from now
+      },
+      forcedDID || threeIdx.id,
+      keyring,
+      threeIdx
+    )
+    return toGeneralJWS(response)
   },
   did_createJWS: async ({ permissions, keyring, threeIdx, origin }, params: CreateJWSParams) => {
     if (!permissions.has(origin)) throw new RPCError(4100, 'Unauthorized')
     // TODO - if the requesting DID is our management key
     // (did:key) we should request explicit permission.
-    let [did, keyFragment] = params.did.split('#') // eslint-disable-line prefer-const
-    let kid, signer
-    if (did.startsWith('did:key:')) {
-      const pubkey = did.split(':')[2]
-      kid = `${did}#${pubkey}`
-      signer = keyring.getMgmtSigner(pubkey)
-    } else {
-      if (did !== threeIdx.id) throw new Error(`Unknown DID: ${did}`)
-      const version = await threeIdx.get3idVersion()
-      if (!keyFragment) keyFragment = keyring.getKeyFragment(version)
-      kid = `${did}?version-id=${version}#${keyFragment}`
-      signer = keyring.getSigner(version)
-    }
-    const header = toStableObject(Object.assign(params.protected || {}, { kid }))
-    const jws = await createJWS(toStableObject(params.payload), signer, header)
-    return { jws }
+    const jws = await sign(params.payload, params.did, keyring, threeIdx, params.protected, params.revocable)
+    return { jws: toGeneralJWS(jws) }
   },
   did_decryptJWE: async ({ permissions, keyring, origin }, params: DecryptJWEParams) => {
     if (!permissions.has(origin)) throw new RPCError(4100, 'Unauthorized')