--- a/src/three-idx.ts
+++ b/src/three-idx.ts
@@ -1,15 +1,15 @@
 import { CeramicApi, Doctype } from '@ceramicnetwork/common'
 import CeramicClient from '@ceramicnetwork/http-client'
-//import { schemas, definitions } from '@ceramicstudio/idx-constants'
+import { schemas, definitions } from '@ceramicstudio/idx-constants'
 import { LinkProof } from '3id-blockchain-utils'
 
 import type { DidProvider } from './did-provider'
 import type { ThreeIdState } from './keyring'
 import type { JWE } from 'did-jwt'
 
-//const KEYCHAIN_DEF = definitions.threeIdKeychain
-const KEYCHAIN_DEF = 'keychain'
-//const { IdentityIndex, ThreeIdKeychain } = schemas
+const KEYCHAIN_DEF = definitions.threeIdKeychain
+const IDX = 'IDX'
+const { IdentityIndex, ThreeIdKeychain } = schemas
 
 export interface EncData {
   jwe?: JWE
@@ -73,15 +73,11 @@ export class ThreeIDX {
 
   async createAuthMapEntry(authEntry: NewAuthEntry): Promise<AuthEntryMap> {
     const authLink = authEntry.linkProof.account
-    // tmp if statement, can be removed when fix released:
-    // https://github.com/ceramicnetwork/js-ceramic/pull/287
-    if (!this.docs[authLink]) {
-      this.docs[authLink] = await this.ceramic.createDocument(
-        'caip10-link',
-        { metadata: { controllers: [authLink] } },
-        { anchor: false, publish: false }
-      )
-    }
+    this.docs[authLink] = await this.ceramic.createDocument(
+      'caip10-link',
+      { metadata: { controllers: [authLink] } },
+      { anchor: false, publish: false }
+    )
     await this.docs[authLink].change({ content: authEntry.linkProof })
     await this.ceramic.pin.add(this.docs[authLink].id)
     const tmpEntry: TempAuthEntry = Object.assign({}, authEntry)
@@ -94,18 +90,15 @@ export class ThreeIDX {
    */
   async createIDX(authEntry?: NewAuthEntry): Promise<void> {
     const entry = authEntry ? await this.createAuthMapEntry(authEntry) : {}
-    await this.createKeychainDoc(entry)
-    this.docs.idx = await this.ceramic.createDocument('tile', {
-      //metadata: { controllers: [this.id], schema: IdentityIndex },
-      metadata: { controllers: [this.id] },
-      content: { [KEYCHAIN_DEF]: this.docs[KEYCHAIN_DEF].id.toUrl('base36') },
-    })
-    await this.docs.threeId.change({
-      content: Object.assign(this.docs.threeId.content, {
-        idx: this.docs.idx.id.baseID.toUrl('base36'),
-      }),
-    })
-    await this.pinAllDocs()
+    await Promise.all([
+      this.loadKeychainDoc(this.id),
+      this.loadIDXDoc(this.id)
+    ])
+    await Promise.all([
+      this.pinAllDocs(),
+      this.updateKeychainDoc(entry),
+      this.addKeychainToIDX()
+    ])
   }
 
   /**
@@ -119,13 +112,13 @@ export class ThreeIDX {
     )
     const did = this.docs[authLink].content
     if (!did) return null
-    this.docs.threeId = await this.ceramic.loadDocument(`ceramic://${did.split(':')[2] as string}`)
-    const idxDocid = this.docs.threeId.content.idx
-    if (!idxDocid) return null
-    this.docs.idx = await this.ceramic.loadDocument(idxDocid)
-    const authKeychainDocid = this.docs.idx.content[KEYCHAIN_DEF]
-    if (!authKeychainDocid) return null
-    this.docs[KEYCHAIN_DEF] = await this.ceramic.loadDocument(authKeychainDocid)
+    await Promise.all([
+      this.loadKeychainDoc(did),
+      this.loadIDXDoc(did),
+      (async () => {
+        this.docs.threeId = await this.ceramic.loadDocument(did.split(':')[2] as string)
+      })()
+    ])
     const linkDocid = this.docs[authLink].id.baseID.toString()
     const { authMap, pastSeeds } = this.docs[KEYCHAIN_DEF].content
     return {
@@ -141,11 +134,11 @@ export class ThreeIDX {
     if (this.docs.idx == null) {
       throw new Error('No IDX doc')
     }
-    await this.docs.idx.change({
-      //metadata: { controllers: [this.id], schema: IdentityIndex },
-      metadata: { controllers: [this.id] },
-      content: { [KEYCHAIN_DEF]: this.docs[KEYCHAIN_DEF].id.toUrl('base36') },
-    })
+    const update = { content: { [KEYCHAIN_DEF]: this.docs[KEYCHAIN_DEF].id.toUrl('base36') } }
+    if (!this.docs.idx.metadata.schema) {
+      update.metadata = { schema: IdentityIndex }
+    }
+    await this.docs.idx.change(update)
   }
 
   /**
@@ -153,17 +146,17 @@ export class ThreeIDX {
    */
   async addAuthEntries(authEntries: Array<NewAuthEntry>): Promise<void> {
     const mapEntries = await Promise.all(authEntries.map(this.createAuthMapEntry.bind(this)))
-    const content = this.docs[KEYCHAIN_DEF].content
+    const { authMap, pastSeeds } = this.docs[KEYCHAIN_DEF].content
     const newAuthEntries = mapEntries.reduce((acc, entry) => Object.assign(acc, entry), {})
-    Object.assign(content.authMap, newAuthEntries)
-    await this.docs[KEYCHAIN_DEF].change({ content })
+    Object.assign(authMap, newAuthEntries)
+    await this.updateKeychainDoc(authMap, pastSeeds)
   }
 
   /**
    * Returns all public keys that is in the auth keychain.
    */
   getAllAuthEntries(): Array<AuthEntry> {
-    if (!this.docs[KEYCHAIN_DEF]) return []
+    if (!this.docs[KEYCHAIN_DEF] || !this.docs[KEYCHAIN_DEF].content.authMap) return []
     const authMap = this.docs[KEYCHAIN_DEF].content.authMap
     return Object.values(authMap)
   }
@@ -176,12 +169,43 @@ export class ThreeIDX {
     )
   }
 
-  async createKeychainDoc(authMap: AuthEntryMap, pastSeeds: Array<JWE> = []): Promise<void> {
-    this.docs[KEYCHAIN_DEF] = await this.ceramic.createDocument('tile', {
-      //metadata: { controllers: [this.id], schema: ThreeIdKeychain },
-      metadata: { controllers: [this.id] },
-      content: { authMap, pastSeeds },
-    })
+  async loadIDXDoc(did: string): Promise<void> {
+    this.docs.idx = await this.ceramic.createDocument('tile',
+      { deterministic: true, metadata: { controllers: [did], family: IDX } },
+      { anchor: false, publish: false }
+    )
+  }
+
+  async addKeychainToIDX(): Promise<void> {
+    const content = this.docs.idx.content
+    if (!content || !content[KEYCHAIN_DEF]) {
+      const update = {
+        content: Object.assign((content || {}), {
+          [KEYCHAIN_DEF]: this.docs[KEYCHAIN_DEF].id.toUrl('base36'),
+        })
+      }
+      if (!this.docs.idx.metadata.schema) {
+        update.metadata = { schema: IdentityIndex }
+      }
+      await this.docs.idx.change(update)
+    }
+  }
+
+  async loadKeychainDoc(did: string): Promise<void> {
+    this.docs[KEYCHAIN_DEF] = await this.ceramic.createDocument('tile',
+      { deterministic: true, metadata: { controllers: [did], family: KEYCHAIN_DEF } },
+      { anchor: false, publish: false }
+    )
+  }
+
+  async updateKeychainDoc(authMap: AuthEntryMap, pastSeeds: Array<JWE> = []): Promise<void> {
+    if (Object.keys(authMap).length !== 0) {
+      const update = { content: { authMap, pastSeeds } }
+      if (!this.docs[KEYCHAIN_DEF].metadata.schema) {
+        update.metadata = { schema: ThreeIdKeychain }
+      }
+      await this.docs[KEYCHAIN_DEF].change(update)
+    }
   }
 
   /**
@@ -209,17 +233,8 @@ export class ThreeIDX {
       if (entry) authMap[link] = entry
       return authMap
     }, {})
-    // Create a new 3ID keychain document
-    await this.createKeychainDoc(authMap, pastSeeds)
+    promises.push(this.updateKeychainDoc(authMap, pastSeeds))
     promises.push(this.pinAllDocs())
-    // Update IDX to point to new keychain
-    promises.push(
-      this.docs.idx.change({
-        content: Object.assign(this.docs.idx.content, {
-          [KEYCHAIN_DEF]: this.docs[KEYCHAIN_DEF].id.toUrl('base36'),
-        }),
-      })
-    )
     await Promise.all(promises)
   }
 }