… package.json
… test/__snapshots__/three-idx.test.js.snap
… test/identity-wallet.test.js
… test/three-idx.test.js