### Changes:
* Update to latest Ceramic packages
* Support DIDs v1 (EIP2844)
* Optimize init time with new IDX structure


### Test running times
As can be seen below we see a **2x** speed improvement.

#### Before this PR
```
 PASS  test/three-idx.test.js (112.012 s)
 PASS  test/identity-wallet.test.js (99.283 s)
 PASS  test/keychain.test.js (23.524 s)
 PASS  test/keyring.test.js (16.938 s)
 PASS  test/did-provider.test.ts
 PASS  test/permissions.test.js
```

### After this PR
```
 PASS  test/three-idx.test.js (53.754 s)
 PASS  test/identity-wallet.test.js (48.873 s)
 PASS  test/keychain.test.js (24.774 s)
 PASS  test/keyring.test.js (14.966 s)
 PASS  test/did-provider.test.ts
 PASS  test/permissions.test.js
```